project1a/adversarial_mnist/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.60.0
